data-augmentation-graphs/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
absl-py=1.3.0=pypi_0
aiohttp=3.8.3=pypi_0
aiosignal=1.2.0=pypi_0
alembic=1.7.7=pypi_0
ase=3.21.1=pypi_0
async-timeout=4.0.2=pypi_0
asynctest=0.13.0=pypi_0
attrs=22.1.0=pypi_0
autopage=0.5.1=pypi_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.12.7=ha878542_0
cached-property=1.5.2=pypi_0
cachetools=4.2.4=pypi_0
certifi=2021.5.30=py36h06a4308_0
chardet=4.0.0=pypi_0
charset-normalizer=2.1.1=pypi_0
cliff=3.10.1=pypi_0
cmaes=0.8.2=pypi_0
cmd2=2.4.2=pypi_0
colorlog=6.7.0=pypi_0
commonmark=0.9.1=pypi_0
cudatoolkit=11.1.1=h6406543_8
cycler=0.10.0=pypi_0
cython=0.29.32=pypi_0
dataclasses=0.8=pyh787bdff_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=pypi_0
einops=0.4.1=pypi_0
entrypoints=0.3=pyhd8ed1ab_1003
expat=2.4.4=h295c915_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h0708190_1
frozenlist=1.2.0=pypi_0
fsspec=2022.1.0=pypi_0
future=0.18.2=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gensim=4.2.0=pypi_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=2.15.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
googledrivedownloader=0.4=pypi_0
graphsampling=0.0.0=dev_0
greenlet=1.1.3=pypi_0
grpcio=1.48.2=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.1.0=pypi_0
icu=58.2=he6710b0_3
idna=2.10=pypi_0
idna-ssl=1.1.0=pypi_0
importlib-metadata=4.8.3=pypi_0
importlib-resources=5.4.0=pypi_0
iopath=0.1.10=pypi_0
ipykernel=5.5.5=py36hcb3619a_0
ipython=5.8.0=py36_1
ipython_genutils=0.2.0=py_1
isodate=0.6.0=pypi_0
jinja2=2.11.3=pypi_0
joblib=1.0.1=pypi_0
jpeg=9b=h024ee3a_2
jupyter-client=7.1.0=pypi_0
jupyter-core=4.9.1=pypi_0
karateclub=1.3.3=pypi_0
keras=2.10.0=pypi_0
kiwisolver=1.3.1=py36h2531618_0
lame=3.100=h7f98852_1001
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
levenshtein=0.20.9=pypi_0
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h2828fa1_18
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libiconv=1.16=h516909a_0
libpng=1.6.37=h21135ba_2
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.3.0=h6de172a_18
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h7f8727e_2
libuv=1.41.0=h7f98852_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.12=h03d6c58_0
littleutils=0.2.2=pypi_0
llvm-openmp=11.0.1=h4bd325d_0
llvmlite=0.35.0=pypi_0
lz4-c=1.9.3=h9c3ff4c_0
mako=1.1.6=pypi_0
markdown=3.3.7=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.4=py36h06a4308_0
matplotlib-base=3.3.4=py36h62a2d02_0
metis=0.2a5=pypi_0
mkl=2020.4=h726a3e6_304
mkl-service=2.3.0=py36h8c4c3a4_2
mkl_fft=1.3.0=py36h92226af_1
mkl_random=1.2.0=py36h7c3b610_1
mtcnn=0.1.1=pypi_0
multidict=5.2.0=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.4=pypi_0
nettle=3.6=he412f7d_0
networkx=2.5.1=pypi_0
networkx-metis=1.0=pypi_0
ninja=1.10.2=h4bd325d_0
numba=0.52.0=pypi_0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
oauthlib=3.2.2=pypi_0
ogb=1.3.4=pypi_0
olefile=0.46=pyh9f0ad1d_1
opencv-python=4.7.0.72=pypi_0
openh264=2.1.1=h780b84a_0
openssl=1.1.1s=h7f8727e_0
optuna=2.10.1=pypi_0
outdated=0.2.0=pypi_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.1.5=pypi_0
pbr=5.10.0=pypi_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.1.2=py36he98fc37_0
pip=21.0.1=py36h06a4308_0
plfit=1.0.3=pypi_0
portalocker=2.7.0=pypi_0
prettytable=2.5.0=pypi_0
prompt_toolkit=1.0.15=py_1
protobuf=3.19.1=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pydeprecate=0.3.1=pypi_0
pygments=2.11.2=pyhd8ed1ab_0
pygsp=0.5.1=pypi_0
pyparsing=2.4.7=pypi_0
pyperclip=1.8.2=pypi_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.13=hdb3f193_0
python-dateutil=2.8.1=pypi_0
python-levenshtein=0.20.9=pypi_0
python-louvain=0.15=pypi_0
python_abi=3.6=1_cp36m
pytorch=1.10.0=py3.6_cuda11.1_cudnn8.0.5_0
pytorch-lightning=1.5.10=pypi_0
pytorch-mutex=1.0=cuda
pytz=2021.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=19.0.2=py36h9947dbf_2
qt=5.9.7=h5867ecd_1
rapidfuzz=2.11.1=pypi_0
rdflib=5.0.0=pypi_0
rdkit-pypi=2021.9.3=pypi_0
readline=8.1=h27cfd23_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=12.6.0=pypi_0
rsa=4.9=pypi_0
scikit-learn=0.24.1=pypi_0
scikit-network=0.20.0=pypi_0
scipy=1.5.4=pypi_0
seaborn=0.11.2=pyhd3eb1b0_0
semseg=0.4.1=dev_0
setuptools=59.5.0=pypi_0
simplegeneric=0.8.1=py_1
sip=4.19.8=py36hf484d3e_0
six=1.15.0=pyh9f0ad1d_0
smart-open=6.3.0=pypi_0
sqlalchemy=1.4.41=pypi_0
sqlite=3.33.0=h62c20be_0
stevedore=3.5.0=pypi_0
tabulate=0.8.10=pypi_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=1.6=pypi_0
termcolor=1.1.0=pypi_0
texttable=1.6.7=pypi_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
torch-cluster=1.5.9=pypi_0
torch-geometric=2.0.3=pypi_0
torch-scatter=2.0.9=pypi_0
torch-sparse=0.6.12=pypi_0
torch-spline-conv=1.2.1=pypi_0
torchaudio=0.10.0=py36_cu111
torchmetrics=0.8.2=pypi_0
torchvision=0.11.0=py36_cu111
tornado=6.1=py36h8f6f2f9_1
tqdm=4.59.0=pypi_0
traitlets=4.3.3=pyhd8ed1ab_2
typing-extensions=4.1.1=pypi_0
urllib3=1.26.3=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
werkzeug=2.0.3=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
yarl=1.7.2=pypi_0
zeromq=4.3.4=h9c3ff4c_0
zipp=3.6.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=ha95c52a_0
